refactor : 공통 클래스 common 패키지로 분리
