+++ src/main/java/com/example/assignment/common/config/QueryDslConfig.java
@@ -1,4 +1,4 @@
-package com.example.assignment.config;
+package com.example.assignment.common.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
