+++ src/main/java/com/example/assignment/common/type/SuccessType.java
@@ -1,4 +1,4 @@
-package com.example.assignment.domain.type;
+package com.example.assignment.common.type;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
